Consolidate API middlewares, move api-keys to device-api
