+++ src/compose/application-manager.ts
@@ -61,8 +61,6 @@ const localModeManager = new LocalModeManager();
 
 export const router = (() => {
 	const $router = express.Router();
-	$router.use(express.urlencoded({ extended: true, limit: '10mb' }));
-	$router.use(express.json({ limit: '10mb' }));
 
 	createV1Api($router);
 	createV2Api($router);
