+++ test/integration/compose/service.spec.ts
@@ -3,7 +3,7 @@ import * as _ from 'lodash';
 import { expect } from 'chai';
 
 import Service from '~/src/compose/service';
-import * as apiKeys from '~/lib/api-keys';
+import * as apiKeys from '~/src/device-api/api-keys';
 
 describe('compose/service: integration tests', () => {
 	describe('Feature labels', () => {
