+++ test/legacy/21-supervisor-api.spec.ts
@@ -7,9 +7,9 @@ import * as apiBinder from '~/src/api-binder';
 import * as deviceState from '~/src/device-state';
 import Log from '~/lib/supervisor-console';
 import SupervisorAPI from '~/src/device-api';
-import * as apiKeys from '~/lib/api-keys';
+import * as apiKeys from '~/src/device-api/api-keys';
 import * as db from '~/src/db';
-import { cloudApiKey } from '~/lib/api-keys';
+import { cloudApiKey } from '~/src/device-api/api-keys';
 
 const mockedOptions = {
 	listenPort: 54321,
